优化倍速播放/增加随机播放
优化倍速播放/增加随机播放

+++ common-c-library/rtmpdump/IDecoder.h
@@ -25,7 +25,7 @@ class VideoDecoderCallback {
   public:
     virtual ~VideoDecoderCallback(){};
     virtual void OnDecodeVideoChangeSize(VideoDecoder *decoder, unsigned int displayWidth, unsigned int displayHeight) = 0;
-    virtual void OnDecodeVideoFrame(VideoDecoder *decoder, void *frame, u_int32_t ts) = 0;
+    virtual void OnDecodeVideoFrame(VideoDecoder *decoder, void *frame, int64_t ts) = 0;
     virtual void OnDecodeVideoError(VideoDecoder *decoder) = 0;
 };
 
@@ -37,8 +37,8 @@ class VideoDecoder {
     virtual bool Reset() = 0;
     virtual void Pause() = 0;
     virtual void ResetStream() = 0;
-    virtual void DecodeVideoKeyFrame(const char *sps, int sps_size, const char *pps, int pps_size, int naluHeaderSize, u_int32_t ts,const char *vps = NULL, int vps_size = 0) = 0;
-    virtual void DecodeVideoFrame(const char *data, int size, u_int32_t dts, u_int32_t pts, VideoFrameType video_type) = 0;
+    virtual void DecodeVideoKeyFrame(const char *sps, int sps_size, const char *pps, int pps_size, int naluHeaderSize, int64_t ts,const char *vps = NULL, int vps_size = 0) = 0;
+    virtual void DecodeVideoFrame(const char *data, int size, int64_t dts, int64_t pts, VideoFrameType video_type) = 0;
     virtual void ReleaseVideoFrame(void *frame) = 0;
     virtual void StartDropFrame() = 0;
     virtual void ClearVideoFrame() = 0;
@@ -48,7 +48,7 @@ class AudioDecoder;
 class AudioDecoderCallback {
   public:
     virtual ~AudioDecoderCallback(){};
-    virtual void OnDecodeAudioFrame(AudioDecoder *decoder, void *frame, u_int32_t ts) = 0;
+    virtual void OnDecodeAudioFrame(AudioDecoder *decoder, void *frame, int64_t ts) = 0;
     virtual void OnDecodeAudioError(AudioDecoder *decoder) = 0;
 };
 
@@ -70,7 +70,7 @@ class AudioDecoder {
         AudioFrameSoundType sound_type,
         const char *data,
         int size,
-        u_int32_t ts) = 0;
+        int64_t ts) = 0;
     virtual void ReleaseAudioFrame(void *frame) = 0;
     virtual void ClearAudioFrame() = 0;
 };
